requirements.txt
python-telegram-bot>=20.0
openai>=0.27.0
nest_asyncio